@@ -35,15 +35,14 @@
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
-    <maven.compiler.source>17</maven.compiler.source>
-    <maven.compiler.target>17</maven.compiler.target>
+    <maven.compiler.release>21</maven.compiler.release>
     <maven.compiler.debuglevel>lines,vars,source</maven.compiler.debuglevel>
 
-    <spring.version>5.3.24</spring.version>
-    <spring.security.version>5.7.5</spring.security.version>
-    <slf4j.version>1.7.30</slf4j.version>
+    <spring.version>6.2.3</spring.version>
+    <spring.security.version>6.4.3</spring.security.version>
+    <slf4j.version>1.7.36</slf4j.version>
     <logback.version>1.2.11</logback.version>
-    <tomcat.version>9.0.62</tomcat.version>
+    <tomcat.version>10.1.36</tomcat.version>
     <junit.jupiter.version>5.8.2</junit.jupiter.version>
     <junit.platform.version>1.8.2</junit.platform.version>
     <junit.vintage.version>5.8.2</junit.vintage.version>
