@@ -7,7 +7,7 @@ import java.util.Queue;
 import java.util.Set;
 import java.util.TreeSet;
 
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 
 import ee.telekom.workflow.graph.Graph;
 import ee.telekom.workflow.graph.GraphValidator;
